+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `parseEnum` utility function
+- `getCurrentQuarter` utility function
 
 ## [2.2.0] - 2025-10-02