+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2026-04-29
+
 ### Added
 
 - `getQuarter` utility function
@@ -201,7 +203,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `eslint-plugin-jsdoc` to lint jsdoc comments
 - `getEnumNameFromValue`,`getEnumValueFromName`, `getEnumNames`,`getEnumNameValues` functions to respectively get the name of an enum from its value, the value from its name, all the names and all the values
 
-[unreleased]: https://github.com/neolution-ch/javascript-utils/compare/2.3.0...HEAD
+[unreleased]: https://github.com/neolution-ch/javascript-utils/compare/2.4.0...HEAD
+[2.4.0]: https://github.com/neolution-ch/javascript-utils/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/neolution-ch/javascript-utils/compare/2.2.0...2.3.0
 [2.2.0]: https://github.com/neolution-ch/javascript-utils/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/neolution-ch/javascript-utils/compare/2.0.0...2.1.0